🛠[회원가입_이메일 인증] : 이메일 컨테이너 생성
